new controller for medication functions & more bug fixes
