+++ server/src/models/Medication.js
@@ -2,11 +2,14 @@ import mongoose from 'mongoose';
 
 const Schema = mongoose.Schema;
 const MedicationSchema = Schema({
+  key: String,
   drugName: String,
   quantity: Number,
   dosage: Number,
   units: String,
-  comments: String
+  comments: String,
+  lastUpdated: Number,
+  needToUpload: Boolean
 });
 
 const MedicationModel = mongoose.model('Medication', MedicationSchema);
